add: trace_ray
