--- apply_mlx.c
@@ -1,622 +1,0 @@
-#include <stdio.h>
-#include <fcntl.h>
-#include "get_next_line.h"
-#include "string_utils.h"
-#include "parser_util_split.h"
-#include <math.h>
-#include "converter.h"
-#include "checker.h"
-#include "mlx.h"
-
-typedef struct s_rgb
-{
-	int	r;
-	int	g;
-	int	b;
-}	t_rgb;
-
-typedef struct s_coordinate
-{
-	double	x;
-	double	y;
-	double	z;
-}	t_coord;
-
-typedef	t_coord	t_vec;
-
-typedef struct s_ambient
-{
-	double	intensity;
-	t_rgb	rgb;
-}	t_ambient;
-
-typedef struct s_light
-{
-	t_coord	coord;
-	double	intensity;
-	t_rgb	rgb;
-}	t_light;
-
-typedef struct s_camera
-{
-	t_coord	coord;
-	t_vec	norm;
-	int		fov;
-}	t_camera;
-
-typedef struct s_sphere
-{
-	t_coord	coord;
-	double	diameter;
-	t_rgb	rgb;
-}	t_sp;
-
-typedef struct s_plane
-{
-	t_coord	coord;
-	t_vec	norm;
-	t_rgb	rgb;
-}	t_pl;
-
-typedef struct s_cylinder
-{
-	t_coord	coord;
-	t_vec	norm;
-	double	diameter;
-	double	height;
-	t_rgb	rgb;
-}	t_cy;
-
-typedef struct s_node
-{
-	void			*data;
-	struct s_node	*next;
-}	t_node;
-
-typedef struct s_rt_info
-{
-	t_ambient	a;
-	t_camera	c;
-	t_light		l;
-	t_node		*sp;
-	t_node		*cy;
-	t_node		*pl;
-}	t_rt_info;
-
-typedef struct s_object
-{
-	int		type;
-	void	*object;
-}	t_object;
-
-int	check_rgb(t_rgb rgb)
-{
-	int	r;
-	int	g;
-	int	b;
-
-	r = rgb.r;
-	g = rgb.g;
-	b = rgb.b;
-	return (-1 < r && r < 256 && -1 < g && g < 256 && -1 < b && b < 256);
-}
-
-int	check_normal(t_vec vec)
-{
-	double	x;
-	double	y;
-	double	z;
-
-	x = vec.x;
-	y = vec.y;
-	z = vec.z;
-	return (-1 <= x && x <= 1 && -1 <= y && y <= 1 && -1 <= z && z <= 1
-		&& sqrt(x * x + y * y + z * z) == 1);
-}
-
-int	free_splitted(char **splitted, int ret)
-{
-	int	i;
-
-	i = -1;
-	while (splitted[++i])
-		free(splitted[i]);
-	free(splitted);
-	return (ret);
-}
-
-int	set_rgb(char *rgb_str, t_rgb *rgb)
-{
-	char	**rgb_info;
-	int		rgb_info_cnt;
-
-	if (!check_comma_cnt(rgb_str))
-		return (0);
-	rgb_info = split_line(rgb_str, ',', &rgb_info_cnt);
-	if (rgb_info_cnt != 3
-		|| !get_int(rgb_info[0], &(rgb->r))
-		|| !get_int(rgb_info[1], &(rgb->g))
-		|| !get_int(rgb_info[2], &(rgb->b)))
-		return (free_splitted(rgb_info, 0));
-	if (!check_rgb(*rgb))
-		return (free_splitted(rgb_info, 0));
-	return (free_splitted(rgb_info, 1));
-}
-
-int	set_coordinate(char *coord_str, t_coord *coord)
-{
-	char	**coord_info;
-	int		coord_info_cnt;
-
-	if (!check_comma_cnt(coord_str))
-		return (0);
-	coord_info = split_line(coord_str, ',', &coord_info_cnt);
-	if (coord_info_cnt != 3
-		|| !get_double(coord_info[0], &(coord->x))
-		|| !get_double(coord_info[1], &(coord->y))
-		|| !get_double(coord_info[2], &(coord->z)))
-		return (free_splitted(coord_info, 0));
-	return (free_splitted(coord_info, 1));
-}
-
-int	set_ambient(char **info, int cnt, t_rt_info *rt_info)
-{
-	double		intensity;
-	t_ambient	*a;
-
-	if (cnt != 3)
-		return (0);
-	a = &(rt_info->a);
-	if (!get_double(info[1], &intensity) || intensity < 0 || 1 < intensity)
-		return (0);
-	if (!set_rgb(info[2], &(a->rgb)))
-		return (0);
-	a->intensity = intensity;
-	return (1);
-}
-
-int	set_light(char **info, int cnt, t_rt_info *rt_info)
-{
-	double	intensitiy;
-	t_light	*l;
-
-	if (cnt != 4)
-		return (0);
-	l = &(rt_info->l);
-	if (!set_coordinate(info[1], &(l->coord)))
-		return (0);
-	if (!get_double(info[2], &intensitiy) || intensitiy < 0 || 1 < intensitiy)
-		return (0);
-	if (!set_rgb(info[3], &(l->rgb)))
-		return (0);
-	l->intensity = intensitiy;
-	return (1);
-}
-
-int	set_camera(char **info, int cnt, t_rt_info *rt_info)
-{
-	int			fov;
-	t_camera	*c;
-
-	if (cnt != 4)
-		return (0);
-	c = &(rt_info->c);
-	if (!set_coordinate(info[1], &(c->coord)))
-		return (0);
-	if (!set_coordinate(info[2], &(c->norm))
-		|| !check_normal(c->norm))
-		return (0);
-	if (!get_int(info[3], &fov)	|| fov < 0 || 180 < fov) // 180 0
-		return (0);
-	c->fov = fov;
-	return (1);
-}
-
-t_node	*get_last_node(t_node *list) // this function is called after malloc, so list is non-null.
-{
-	while (list->next)
-		list = list->next;
-	return (list);
-}
-
-int	set_plane(char **info, int cnt, t_rt_info *rt_info)
-{
-	t_pl	*pl;
-
-	if (cnt != 4)
-		return (0);
-	pl = (t_pl *)(get_last_node(rt_info->pl)->data);
-	if (!set_coordinate(info[1], &(pl->coord)))
-		return (0);
-	if (!set_coordinate(info[2], &(pl->norm))
-		|| !check_normal(pl->norm))
-		return (0);
-	if (!set_rgb(info[3], &(pl->rgb)))
-		return (0);
-	return (1);
-}
-
-int	set_sphere(char **info, int cnt, t_rt_info *rt_info)
-{
-	double		diameter;
-	t_sp	*sp;
-
-	if (cnt != 4)
-		return (0);
-	sp = (t_sp *)(get_last_node(rt_info->sp)->data);
-	if (!set_coordinate(info[1], &(sp->coord)))
-		return (0);
-	if (!get_double(info[2], &diameter) || diameter < 0)
-		return (0);
-	if (!set_rgb(info[3], &(sp->rgb)))
-		return (0);
-	sp->diameter = diameter;
-	return (1);
-}
-
-int	set_cylinder(char **info, int cnt, t_rt_info *rt_info)
-{
-	double		diameter;
-	double		height;
-	t_cy	*cy;
-
-	if (cnt != 6)
-		return (0);
-	cy = (t_cy *)(get_last_node(rt_info->cy)->data);
-	if (!set_coordinate(info[1], &(cy->coord)))
-		return (0);
-	if (!set_coordinate(info[2], &(cy->norm))
-		|| !check_normal(cy->norm))
-		return (0);
-	if (!get_double(info[3], &diameter) || diameter < 0)
-		return (0);
-	if (!get_double(info[4], &height) || height < 0)
-		return (0);
-	if (!set_rgb(info[5], &(cy->rgb)))
-	cy->diameter = diameter;
-	cy->height = height;
-	return (1);
-}
-
-int	alloc_sphere(void **ptr){
-	t_sp	*sp;
-
-	sp = (t_sp *)malloc(sizeof(t_sp));
-	if (!sp)
-		return (0);
-	*ptr = (void *)sp;
-	return (1);
-}
-
-int	alloc_plane(void **ptr){
-	t_pl	*pl;
-
-	pl = (t_pl *)malloc(sizeof(t_pl));
-	if (!pl)
-		return (0);
-	*ptr = (void *)pl;
-	return (1);
-}
-
-int	alloc_cylinder(void **ptr){
-	t_cy	*cy;
-
-	cy = (t_cy *)malloc(sizeof(t_cy));
-	if (!cy)
-		return (0);
-	*ptr = (void *)cy;
-	return (1);
-}
-
-int	alloc_new_node(t_node **node, int index)
-{
-	static int	(*allocator[3])(void **)
-		= {alloc_sphere, alloc_plane, alloc_cylinder};
-
-	*node = (t_node *)malloc(sizeof(t_node));
-	if (!(*node))
-		return (0);
-	(*node)->next = 0;
-	if (!(*allocator[index - 3])(&((*node)->data)))
-	{
-		free(*node);
-		return (0);
-	}
-	return (1);
-}
-
-void	append_node(t_node **lst, t_node *new)
-{
-	if (!(*lst))
-	{
-		*lst = new;
-		return ;
-	}
-	while ((*lst)->next)
-		lst = &((*lst)->next);
-	(*lst)->next = new;
-}
-
-void	clear_list(t_node **lst) // will be used in error handling function right before exit().
-{
-	t_node	*next;
-
-	while (*lst)
-	{
-		next = (*lst)->next;
-		free((*lst)->data);
-		free(*lst);
-		*lst = next;
-	}
-	*lst = 0;
-}
-
-int	set_object_list(t_rt_info *rt, int index)
-{
-	t_node	**objects;
-	t_node	*new_node;
-
-	if (!alloc_new_node(&new_node, index))
-		return (0);
-	objects = &(rt->sp);
-	if (index == 4)
-		objects = &(rt->pl);
-	else if (index == 5)
-		objects = &(rt->cy);
-	append_node(objects, new_node);
-	return (1);
-}
-
-/*
-splitted
-	1. 널 포인터 및 빈 문자열일 때 : 가드가 거름.
-	2. 내용은 있지만, 유효하지 않은 상태일 때 : 각 set_* 함수들이 걸러야 함.
-*/
-int	set_rt_info(char *line, t_rt_info *rt, int *mask)
-{
-	char		**splitted;
-	int			cnt;
-	int			index;
-	static int	(*fp[6])(char **, int, t_rt_info *) = {set_ambient, set_camera,
-		set_light, set_sphere, set_plane, set_cylinder};
-	
-	splitted = split_line(line, ' ', &cnt); // all split_lines need null-guard.
-	if (!(splitted && splitted[0]))
-		return (free_splitted(splitted, 0));
-	index = check_identifier(splitted[0]);
-	if (index == -1)
-		return (free_splitted(splitted, 0)); // caller must print error if set_rt_info returns 0.
-	if (index < 3 && (*mask & 1 << index || !(*fp[index])(splitted, cnt, rt))) // A, C, L shouldn't be on mutiple lines and don't need malloc.
-		return (free_splitted(splitted, 0));
-	*mask |= 1 << index;
-	if (2 < index
-		&& (!set_object_list(rt, index)	|| !(*fp[index])(splitted, cnt, rt)))
-		return (free_splitted(splitted, 0));
-	return (free_splitted(splitted, 1));
-}
-
-t_vec	vec_add(t_vec v1, t_vec v2)
-{
-	t_vec	ret;
-
-	ret.x = v1.x + v2.x;
-	ret.y = v1.y + v2.y;
-	ret.z = v1.z + v2.z;
-	return (ret);
-}
-
-t_vec	vec_sub(t_vec v1, t_vec v2)
-{
-	t_vec	ret;
-
-	ret.x = v1.x - v2.x;
-	ret.y = v1.y - v2.y;
-	ret.z = v1.z - v2.z;
-	return (ret);
-}
-
-t_vec	vec_scale(t_vec vec, double scalar)
-{
-	vec.x *= scalar;
-	vec.y *= scalar;
-	vec.z *= scalar;
-	return (vec);
-}
-
-double	vec_dot(t_vec v1, t_vec v2)
-{
-	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z);
-}
-
-t_vec	vec_cross(t_vec v1, t_vec v2)
-{
-	t_vec	ret;
-
-	ret.x = v1.y * v2.z - v1.z * v2.y;
-	ret.y = v1.z * v2.x - v1.x * v2.z;
-	ret.z = v1.x * v2.y - v1.y * v2.x;
-	return (ret);
-}
-
-double	vec_magnitude(t_vec vec)
-{
-	return (sqrt(vec_dot(vec, vec)));
-}
-
-int	open_file(char *path)
-{
-	int	fd;
-	int	len;
-
-	fd = -1;
-	len = ft_strlen(path);
-	if (path[len - 3] == '.' && path[len - 2] == 'r' && path[len - 1] == 't')
-		fd = open(path, O_RDONLY); // is read-only enough?
-	// call error handling function to exit.
-	return (fd);
-}
-
-int	read_file(int fd, t_rt_info *rt_info)
-{
-	char		*line;
-	int			flag;
-	static int	mask = 0; // bit-mask to check A, C, L should be declared only and at least once.
-
-	flag = 1;
-	while (flag)
-	{
-		line = get_next_line(fd);
-		if (line == 0)
-			break ;
-		if (line[0] != '\0')
-			flag = set_rt_info(line, rt_info, &mask);
-		free(line);
-	}
-	if (!flag || !(mask & 1 << 0 && mask & 1 << 1 && mask & 1 << 2))
-		return (0);
-	return (1);
-}
-///////////////////// mlx
-#define ESC_KEY		53
-#define LEFT			123
-#define UP				126
-#define RIGHT			124
-#define DOWN			125
-#define P_WID 1280
-#define P_HEI 720
-
-typedef struct s_img
-{
-	void	*ptr;
-	char	*addr;
-	int		bits_per_pixel;
-	int		line_length;
-	int		endian;
-}	t_img;
-
-typedef struct s_vars
-{
-	void		*mlx;
-	void		*win;
-	t_img		img;
-}	t_vars;
-
-int	init_mlx_pointers(t_vars *vars)
-{
-	vars->mlx = mlx_init();
-	if (!vars->mlx)
-		return (0);
-	vars->win = mlx_new_window(vars->mlx, P_WID, P_HEI, "miniRT");
-	if (!vars->win)
-		return (0);
-	vars->img.ptr = mlx_new_image(vars->mlx, P_WID, P_HEI);
-	if (!vars->img.ptr)
-		return (0);
-	vars->img.addr = mlx_get_data_addr(vars->img.ptr, &vars->img.bits_per_pixel,
-			&vars->img.line_length, &vars->img.endian);
-	if (!vars->img.addr)
-		return (0);
-	return (1);
-}
-
-
-
-///////////////////// get_ray
-t_vec	get_ab_vec(t_vec v)
-{
-	t_vec	ret;
-
-	if (v.x == 0 && v.y == 0 && (v.z == 1 || v.z == -1))
-	{
-		ret.x = 1;
-		ret.y = 0;
-		ret.z = 0;
-	}
-	else
-	{
-		ret.x = 0;
-		ret.y = 0;
-		ret.z = 1;
-	}
-	return (ret);
-}
-
-void	get_pixel_info(t_camera c, t_vec *p_h, t_vec *p_v,\
-		t_coord *top_left)
-{
-	t_vec	h;
-	t_vec	v;
-	t_vec	tmp;
-	double		vp_h;
-	double		vp_w;
-
-	vp_w = tan((c.fov * M_PI / 180.0) / 2.0) * 2;
-	vp_h = vp_w * ((double)P_HEI / (double)P_WID);
-	h = vec_cross(c.norm, get_ab_vec(c.norm));
-	v = vec_cross(c.norm, h);
-	h = vec_scale(h, (double)1 / vec_magnitude(h) * (vp_w / (double)P_WID));
-	v = vec_scale(v, (double)1 / vec_magnitude(v) * (vp_h / (double)P_HEI));
-	*p_h = h;
-	*p_v = v;
-	tmp = vec_add(c.coord, c.norm);
-	tmp = vec_sub(tmp, vec_scale(h, (double)P_WID / (double)2));
-	tmp = vec_sub(tmp, vec_scale(v, (double)P_HEI / (double)2));
-	*top_left = tmp;
-}
-
-void	print_vec(t_vec v)	//function for test.
-{
-	printf("(%lf, %lf, %lf)", v.x, v.y, v.z);
-}
-
-int draw_img(t_rt_info *rt_info, t_vars *vars)
-{
-	t_coord	top_left;
-	t_vec		p_h;
-	t_vec		p_v;
-	int				i;
-	int				j;
-
-	get_pixel_info(rt_info->c, &p_h, &p_v, &top_left);
-	i = -1;
-	while (++i < P_HEI)
-	{
-		j = -1;
-		while (++j < P_WID)
-		{
-			print_vec(vec_sub(vec_add(vec_scale(p_v, j), \
-			vec_add(top_left, vec_scale(p_h, i))), rt_info->c.coord));	// for test.
-			printf(" ");
-		}
-		printf("\n");
-	}
-	mlx_put_image_to_window(vars->mlx, vars->win, vars->img.ptr, 0, 0);
-	return (1);
-}
-////////////////////////
-
-int	main(int argc, char **argv)
-{
-	int			fd;
-	t_rt_info	rt_info;
-	t_vars		vars;
-
-	if (argc != 2)
-		return (0); // call error handling function with proper error message.
-	init_mlx_pointers(&vars);
-	fd = open_file(argv[1]);
-	if (fd < 0) // remove this when error handling function is completed.
-		return (0); // 에러 문자열 출력하고 처리해주기
-	rt_info.sp = 0;
-	rt_info.pl = 0;
-	rt_info.cy = 0;
-	if (read_file(fd, &rt_info))
-		printf("%s\n", "valid format");
-	else
-		printf("%s\n", "invalid format");
-	clear_list(&(rt_info.sp));
-	clear_list(&(rt_info.pl));
-	clear_list(&(rt_info.cy));
-	mlx_loop(vars.mlx);
-	return (0);
-}
